Rename FlyDeletionTool to FlySelectionTool and update related functionality; add camera frame selection icon and localizations for fly selection tool.
