+++ src/ui/bottom-toolbar.ts
@@ -9,7 +9,7 @@ import pickerSvg from './svg/select-picker.svg';
 import polygonSvg from './svg/select-poly.svg';
 import sphereSvg from './svg/select-sphere.svg';
 import undoSvg from './svg/undo.svg';
-import trashSvg from './svg/trash.svg'; // Import the new trash icon SVG
+import cameraFrameSelectionSvg from './svg/camera-frame-selection.svg'; // Import the camera frame selection icon
 import { Tooltips } from './tooltips';
 // import cropSvg from './svg/crop.svg';
 
@@ -58,8 +58,8 @@ class BottomToolbar extends Container {
             class: 'bottom-toolbar-tool'
         });
 
-        const flyDelete = new Button({
-            id: 'bottom-toolbar-fly-delete',
+        const flySelect = new Button({ // Renamed from flyDelete
+            id: 'bottom-toolbar-fly-select', // Renamed ID
             class: 'bottom-toolbar-tool'
         });
 
@@ -113,7 +113,7 @@ class BottomToolbar extends Container {
         picker.dom.appendChild(createSvg(pickerSvg));
         polygon.dom.appendChild(createSvg(polygonSvg));
         brush.dom.appendChild(createSvg(brushSvg));
-        flyDelete.dom.appendChild(createSvg(trashSvg)); // Use the imported SVG
+        flySelect.dom.appendChild(createSvg(cameraFrameSelectionSvg)); // Use the imported SVG for flySelect
         sphere.dom.appendChild(createSvg(sphereSvg));
         lasso.dom.appendChild(createSvg(lassoSvg));
         // crop.dom.appendChild(createSvg(cropSvg));
@@ -126,7 +126,7 @@ class BottomToolbar extends Container {
         this.append(polygon);
         this.append(brush);
         this.append(new Element({ class: 'bottom-toolbar-separator' }));
-        this.append(flyDelete); // Add the new button here
+        this.append(flySelect); // Add the renamed button
         this.append(sphere);
         // this.append(crop);
         this.append(new Element({ class: 'bottom-toolbar-separator' }));
@@ -141,7 +141,7 @@ class BottomToolbar extends Container {
         polygon.dom.addEventListener('click', () => events.fire('tool.polygonSelection'));
         lasso.dom.addEventListener('click', () => events.fire('tool.lassoSelection'));
         brush.dom.addEventListener('click', () => events.fire('tool.brushSelection'));
-        flyDelete.dom.addEventListener('click', () => events.fire('tool.flyDeletion')); // Fire new event
+        flySelect.dom.addEventListener('click', () => events.fire('tool.flySelection')); // Fire new event
         picker.dom.addEventListener('click', () => events.fire('tool.rectSelection'));
         sphere.dom.addEventListener('click', () => events.fire('tool.sphereSelection'));
         translate.dom.addEventListener('click', () => events.fire('tool.move'));
@@ -160,7 +160,7 @@ class BottomToolbar extends Container {
         events.on('tool.activated', (toolName: string) => {
             picker.class[toolName === 'rectSelection' ? 'add' : 'remove']('active');
             brush.class[toolName === 'brushSelection' ? 'add' : 'remove']('active');
-            flyDelete.class[toolName === 'flyDeletion' ? 'add' : 'remove']('active'); // Handle active state
+            flySelect.class[toolName === 'flySelection' ? 'add' : 'remove']('active'); // Handle active state for flySelect
             polygon.class[toolName === 'polygonSelection' ? 'add' : 'remove']('active');
             lasso.class[toolName === 'lassoSelection' ? 'add' : 'remove']('active');
             sphere.class[toolName === 'sphereSelection' ? 'add' : 'remove']('active');
@@ -182,7 +182,7 @@ class BottomToolbar extends Container {
         tooltips.register(redo, localize('tooltip.redo'));
         tooltips.register(picker, localize('tooltip.picker'));
         tooltips.register(brush, localize('tooltip.brush'));
-        tooltips.register(flyDelete, localize('tooltip.fly-delete'));
+        tooltips.register(flySelect, localize('tooltip.fly-select')); // Use new tooltip key
         tooltips.register(polygon, localize('tooltip.polygon'));
         tooltips.register(lasso, 'Lasso Select');
         tooltips.register(sphere, localize('tooltip.sphere'));
