+++ src/ui/localization.ts
@@ -175,6 +175,7 @@ const localizeInit = () => {
                     'tooltip.brush': 'Pinselselektion ( B )',
                     'tooltip.sphere': 'Kugelselektion',
                     'tooltip.fly-delete': 'Fluglöschwerkzeug', // Added
+                    'tooltip.fly-select': 'Flugauswahl', // Added tooltip for fly select
                     'tooltip.translate': 'Verschieben ( 1 )',
                     'tooltip.rotate': 'Drehen ( 2 )',
                     'tooltip.scale': 'Skalieren ( 3 )',
@@ -416,6 +417,7 @@ const localizeInit = () => {
                     'tooltip.brush': 'Brush Select ( B )',
                     'tooltip.sphere': 'Sphere Select',
                     'tooltip.fly-delete': 'Fly Deletion Tool', // Added
+                    'tooltip.fly-select': 'Fly Select', // Added tooltip for fly select
                     'tooltip.translate': 'Translate ( 1 )',
                     'tooltip.rotate': 'Rotate ( 2 )',
                     'tooltip.scale': 'Scale ( 3 )',
@@ -646,6 +648,7 @@ const localizeInit = () => {
                     'tooltip.brush': 'Sélection avec pinceau ( B )',
                     'tooltip.sphere': 'Sélection avec sphère',
                     'tooltip.fly-delete': 'Outil de suppression en vol', // Added
+                    'tooltip.fly-select': 'Sélection en vol', // Added tooltip for fly select
                     'tooltip.translate': 'Translation ( 1 )',
                     'tooltip.rotate': 'Rotation ( 2 )',
                     'tooltip.scale': 'Échelle ( 3 )',
@@ -872,6 +875,7 @@ const localizeInit = () => {
                     'tooltip.brush': 'ブラシ選択 ( B )',
                     'tooltip.sphere': '球で選択',
                     'tooltip.fly-delete': '飛行削除ツール', // Added
+                    'tooltip.fly-select': '飛行選択', // Added tooltip for fly select
                     'tooltip.translate': '移動 ( 1 )',
                     'tooltip.rotate': '回転 ( 2 )',
                     'tooltip.scale': 'スケール ( 3 )',
@@ -1097,6 +1101,7 @@ const localizeInit = () => {
                     'tooltip.brush': '브러시 선택 ( B )',
                     'tooltip.sphere': '구 선택',
                     'tooltip.fly-delete': '비행 삭제 도구', // Added
+                    'tooltip.fly-select': '비행 선택', // Added tooltip for fly select
                     'tooltip.translate': '이동 ( 1 )',
                     'tooltip.rotate': '회전 ( 2 )',
                     'tooltip.scale': '크기 조정 ( 3 )',
@@ -1323,6 +1328,7 @@ const localizeInit = () => {
                     'tooltip.brush': '画笔 ( B )',
                     'tooltip.sphere': '球选择',
                     'tooltip.fly-delete': '飞行删除工具', // Added
+                    'tooltip.fly-select': '飞行选择', // Added tooltip for fly select
                     'tooltip.translate': '移动 ( 1 )',
                     'tooltip.rotate': '旋转 ( 2 )',
                     'tooltip.scale': '缩放 ( 3 )',
